# gin-swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "liuxing test helloworld",
        "title": "helloworld",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "lx",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "lx.helloworld.com",
    "basePath": "/base/path",
    "paths": {
        "/api/v1/user/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "禅道登录接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户名",
                        "name": "account",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/main.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.StandardResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "禅道登出接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "session_id",
                        "name": "zentaosid",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/main.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "请求错误",
                        "schema": {
                            "$ref": "#/definitions/main.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.StandardResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}